requirements.txt

fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
spacy
python-dateutil
pydantic
